load efps

+++ calc_metrics.py
@@ -88,34 +88,38 @@ for i in range(start_idx, 4001, 5):
     G.load_state_dict(torch.load(f"{models_dir}/G_{i}.pt", map_location="cuda"))
     G.eval()
 
-    gen_jets = train.gen_multi_batch(
-        {"embed_dim": model_args.gapt_embed_dim},
-        G,
-        2048,
-        50000,
-        30,
-        model="gapt",
-        out_device="cpu",
-        labels=real_jf[:50000],
-        detach=True,
-    )
+    if os.path.exists(f"{jets_dir}/{i}_gen_efps.npy"):
+        gen_efps = np.load(f"{jets_dir}/{i}_gen_efps.npy")
 
-    gen_jets = jetnet.utils.gen_jet_corrections(
-        particle_norm(gen_jets, inverse=True),
-        ret_mask_separate=True,
-        zero_mask_particles=True,
-    )
+    else:
+        gen_jets = train.gen_multi_batch(
+            {"embed_dim": model_args.gapt_embed_dim},
+            G,
+            2048,
+            50000,
+            30,
+            model="gapt",
+            out_device="cpu",
+            labels=real_jf[:50000],
+            detach=True,
+        )
 
-    gen_mask = gen_jets[1]
-    gen_jets = gen_jets[0]
+        gen_jets = jetnet.utils.gen_jet_corrections(
+            particle_norm(gen_jets, inverse=True),
+            ret_mask_separate=True,
+            zero_mask_particles=True,
+        )
 
-    gen_mask = gen_mask.numpy()
-    gen_jets = gen_jets.numpy()
+        gen_mask = gen_jets[1]
+        gen_jets = gen_jets[0]
 
-    gen_efps = jetnet.utils.efps(gen_jets, efpset_args=[("d<=", 4)], efp_jobs=6)
+        gen_mask = gen_mask.numpy()
+        gen_jets = gen_jets.numpy()
 
-    np.save(f"{jets_dir}/{i}_gen_jets.npy", gen_jets)
-    np.save(f"{jets_dir}/{i}_gen_efps.npy", gen_efps)
+        gen_efps = jetnet.utils.efps(gen_jets, efpset_args=[("d<=", 4)], efp_jobs=6)
+
+        np.save(f"{jets_dir}/{i}_gen_jets.npy", gen_jets)
+        np.save(f"{jets_dir}/{i}_gen_efps.npy", gen_efps)
 
     print(f"{datetime.datetime.now()} Calculating KPD")
